:sparkles: delete table have fixed.

+++ backend/src/table-1/table-1.service.ts
@@ -63,6 +63,9 @@ export class Table1Service {
         try {
             const eye_photo = await this.eyePhotosService.getEyePhotoById(eye_photo_id)
             const table = await this.table1Repository.findOne({ where: { eye_photo } })
+            if(!table) {
+                return "success"
+            }
             await this.table1Repository.delete(table.table_1_id)
             return "success"
         } catch(e) {
