+++ backend/src/table-13/table-13.service.ts
@@ -112,6 +112,9 @@ export class Table13Service {
         try {
             const eye_photo = await this.eyePhotosService.getEyePhotoById(eye_photo_id)
             const table = await this.table13Repository.findOne({ where: { eye_photo } })
+            if(!table) {
+                return "success"
+            }
             await this.table13Repository.delete(table.table_13_id)
             return "success"
         } catch(e) {
