+++ backend/src/table-5/table-5.service.ts
@@ -65,6 +65,9 @@ export class Table5Service {
         try {
             const eye_photo = await this.eyePhotosService.getEyePhotoById(eye_photo_id)
             const table = await this.table5Repository.findOne({ where: { eye_photo } })
+            if(!table) {
+                return "success"
+            }
             await this.table5Repository.delete(table.table_5_id)
             return "success"
         } catch(e) {
